requirements.txt
requests
pystray
pillow
schedule
win10toast-click
pywin32
